+++ sdk/keyvault/azsecrets/client.go
@@ -8,21 +8,18 @@ package azsecrets
 
 import (
 	"context"
-	"encoding/json"
-	"errors"
 	"net/http"
-	"time"
 
 	"github.com/Azure/azure-sdk-for-go/sdk/azcore"
 	"github.com/Azure/azure-sdk-for-go/sdk/azcore/policy"
 	"github.com/Azure/azure-sdk-for-go/sdk/azcore/runtime"
-	"github.com/Azure/azure-sdk-for-go/sdk/keyvault/azsecrets/internal"
+	"github.com/Azure/azure-sdk-for-go/sdk/keyvault/azsecrets/internal/generated"
 	shared "github.com/Azure/azure-sdk-for-go/sdk/keyvault/internal"
 )
 
 // Client interacts with Key Vault secrets.
 type Client struct {
-	kvClient *internal.KeyVaultClient
+	kvClient *generated.KeyVaultClient
 	vaultUrl string
 }
 
@@ -31,36 +28,17 @@ type ClientOptions struct {
 	azcore.ClientOptions
 }
 
-func (c *ClientOptions) toConnectionOptions() *policy.ClientOptions {
-	if c == nil {
-		return nil
-	}
-
-	return &policy.ClientOptions{
-		Logging:          c.Logging,
-		Retry:            c.Retry,
-		Telemetry:        c.Telemetry,
-		Transport:        c.Transport,
-		PerCallPolicies:  c.PerCallPolicies,
-		PerRetryPolicies: c.PerRetryPolicies,
-	}
-}
-
 // NewClient constructs a Client that accesses a Key Vault's secrets.
 func NewClient(vaultURL string, credential azcore.TokenCredential, options *ClientOptions) (*Client, error) {
 	if options == nil {
 		options = &ClientOptions{}
 	}
-
-	conOptions := options.toConnectionOptions()
-
-	conOptions.PerRetryPolicies = append(
-		conOptions.PerRetryPolicies,
-		shared.NewKeyVaultChallengePolicy(credential),
-	)
-
+	plOpts := runtime.PipelineOptions{
+		PerRetry: []policy.Policy{shared.NewKeyVaultChallengePolicy(credential)},
+	}
+	pl := runtime.NewPipeline(moduleName, version, plOpts, &options.ClientOptions)
 	return &Client{
-		kvClient: internal.NewKeyVaultClient(conOptions),
+		kvClient: generated.NewKeyVaultClient(pl),
 		vaultUrl: vaultURL,
 	}, nil
 }
@@ -78,11 +56,11 @@ type GetSecretOptions struct {
 }
 
 // convert the exposed options struct to the internal one.
-func (g *GetSecretOptions) toGenerated() *internal.KeyVaultClientGetSecretOptions {
+func (g *GetSecretOptions) toGenerated() *generated.KeyVaultClientGetSecretOptions {
 	if g == nil {
-		return &internal.KeyVaultClientGetSecretOptions{}
+		return &generated.KeyVaultClientGetSecretOptions{}
 	}
-	return &internal.KeyVaultClientGetSecretOptions{}
+	return &generated.KeyVaultClientGetSecretOptions{}
 }
 
 // GetSecretResponse is returned by GetSecret.
@@ -90,7 +68,7 @@ type GetSecretResponse struct {
 	Secret
 }
 
-func getSecretResponseFromGenerated(i internal.KeyVaultClientGetSecretResponse) GetSecretResponse {
+func getSecretResponseFromGenerated(i generated.KeyVaultClientGetSecretResponse) GetSecretResponse {
 	vaultURL, name, version := shared.ParseID(i.ID)
 	return GetSecretResponse{
 		Secret: Secret{
@@ -142,11 +120,11 @@ type SetSecretOptions struct {
 }
 
 // Convert the exposed struct to the generated code version
-func (s *SetSecretOptions) toGenerated() *internal.KeyVaultClientSetSecretOptions {
+func (s *SetSecretOptions) toGenerated() *generated.KeyVaultClientSetSecretOptions {
 	if s == nil {
 		return nil
 	}
-	return &internal.KeyVaultClientSetSecretOptions{}
+	return &generated.KeyVaultClientSetSecretOptions{}
 }
 
 // SetSecretResponse is returned by SetSecret.
@@ -155,7 +133,7 @@ type SetSecretResponse struct {
 }
 
 // convert generated response to publicly exposed response.
-func setSecretResponseFromGenerated(i internal.KeyVaultClientSetSecretResponse) SetSecretResponse {
+func setSecretResponseFromGenerated(i generated.KeyVaultClientSetSecretResponse) SetSecretResponse {
 	vaultURL, name, version := shared.ParseID(i.ID)
 	return SetSecretResponse{
 		Secret: Secret{
@@ -187,11 +165,11 @@ func (c *Client) SetSecret(ctx context.Context, name string, value string, optio
 	if options == nil {
 		options = &SetSecretOptions{}
 	}
-	var secretAttribs internal.SecretAttributes
+	var secretAttribs generated.SecretAttributes
 	if options.Properties != nil {
 		secretAttribs = *options.Properties.toGenerated()
 	}
-	resp, err := c.kvClient.SetSecret(ctx, c.vaultUrl, name, internal.SecretSetParameters{
+	resp, err := c.kvClient.SetSecret(ctx, c.vaultUrl, name, generated.SecretSetParameters{
 		Value:            &value,
 		ContentType:      options.ContentType,
 		SecretAttributes: &secretAttribs,
@@ -208,7 +186,7 @@ type DeleteSecretResponse struct {
 	DeletedSecret
 }
 
-func deleteSecretResponseFromGenerated(i internal.KeyVaultClientDeleteSecretResponse) DeleteSecretResponse {
+func deleteSecretResponseFromGenerated(i generated.KeyVaultClientDeleteSecretResponse) DeleteSecretResponse {
 	vaultURL, name, version := shared.ParseID(i.ID)
 	return DeleteSecretResponse{
 		DeletedSecret: DeletedSecret{
@@ -240,122 +218,44 @@ func deleteSecretResponseFromGenerated(i internal.KeyVaultClientDeleteSecretResp
 // BeginDeleteSecretOptions contains optional parameters for BeginDeleteSecret.
 type BeginDeleteSecretOptions struct {
 	// ResumeToken is a string to rehydrate a poller for an operation that has already begun.
-	ResumeToken *string
+	ResumeToken string
 }
 
 // convert public options to generated options struct
-func (b *BeginDeleteSecretOptions) toGenerated() *internal.KeyVaultClientDeleteSecretOptions {
-	return &internal.KeyVaultClientDeleteSecretOptions{}
-}
-
-// DeleteSecretPoller is returned by BeginDeleteSecret.
-type DeleteSecretPoller struct {
-	secretName     string // This is the secret to Poll for in GetDeletedSecret
-	vaultUrl       string
-	client         *internal.KeyVaultClient
-	deleteResponse internal.KeyVaultClientDeleteSecretResponse
-	lastResponse   internal.KeyVaultClientGetDeletedSecretResponse
-	rawResponse    *http.Response
-	resumeToken    string
-}
-
-// Done returns true if the LRO has reached a terminal state
-func (s *DeleteSecretPoller) Done() bool {
-	if s.rawResponse == nil {
-		return false
-	}
-	return s.rawResponse.StatusCode == http.StatusOK
-}
-
-// Poll fetches the latest state of the delete operation.
-func (s *DeleteSecretPoller) Poll(ctx context.Context) (*http.Response, error) {
-	var rawResp *http.Response
-	ctx = runtime.WithCaptureResponse(ctx, &rawResp)
-	resp, err := s.client.GetDeletedSecret(ctx, s.vaultUrl, s.secretName, nil)
-	if err == nil {
-		// Service recognizes DeletedSecret, operation is done
-		s.lastResponse = resp
-		s.rawResponse = rawResp
-		return rawResp, nil
-	}
-	if rawResp != nil && rawResp.StatusCode == http.StatusNotFound {
-		// This is the expected result
-		s.rawResponse = rawResp
-		return rawResp, nil
-	}
-	return rawResp, err
-}
-
-// FinalResponse returns the final response after the secret is deleted.
-func (s *DeleteSecretPoller) FinalResponse(ctx context.Context) (DeleteSecretResponse, error) {
-	return deleteSecretResponseFromGenerated(s.deleteResponse), nil
-}
-
-// PollUntilDone polls Key Vault until the secret is deleted. The t parameter determines the wait between polls.
-func (s *DeleteSecretPoller) PollUntilDone(ctx context.Context, t time.Duration) (DeleteSecretResponse, error) {
-	for {
-		resp, err := s.Poll(ctx)
-		if err != nil {
-			return DeleteSecretResponse{}, err
-		}
-		s.rawResponse = resp
-		if s.Done() {
-			break
-		}
-		time.Sleep(t)
-	}
-	return deleteSecretResponseFromGenerated(s.deleteResponse), nil
-}
-
-// ResumeToken returns a token for resuming polling at a later time
-func (s *DeleteSecretPoller) ResumeToken() (string, error) {
-	return s.resumeToken, nil
+func (b *BeginDeleteSecretOptions) toGenerated() *generated.KeyVaultClientDeleteSecretOptions {
+	return &generated.KeyVaultClientDeleteSecretOptions{}
 }
 
 // BeginDeleteSecret deletes all versions of a secret. It returns a Poller that enables waiting for Key Vault to finish
 // deleting the secret.
-func (c *Client) BeginDeleteSecret(ctx context.Context, name string, options *BeginDeleteSecretOptions) (*DeleteSecretPoller, error) {
+func (c *Client) BeginDeleteSecret(ctx context.Context, name string, options *BeginDeleteSecretOptions) (*runtime.Poller[DeleteSecretResponse], error) {
 	if options == nil {
 		options = &BeginDeleteSecretOptions{}
 	}
-	var resumeToken string
-	var delResp internal.KeyVaultClientDeleteSecretResponse
-	var err error
-	if options.ResumeToken == nil {
-		delResp, err = c.kvClient.DeleteSecret(ctx, c.vaultUrl, name, options.toGenerated())
-		if err != nil {
-			return nil, err
-		}
 
-		marshalled, err := json.Marshal(delResp)
-		if err != nil {
-			return nil, err
-		}
-		resumeToken = string(marshalled)
-	} else {
-		resumeToken = *options.ResumeToken
-		err = json.Unmarshal([]byte(resumeToken), &delResp)
-		if err != nil {
-			return nil, err
-		}
+	handler := beginDeleteSecretOperation{
+		poll: func(ctx context.Context) (*http.Response, error) {
+			req, err := c.kvClient.GetDeletedSecretCreateRequest(ctx, c.vaultUrl, name, nil)
+			if err != nil {
+				return nil, err
+			}
+			return c.kvClient.Pipeline().Do(req)
+		},
 	}
 
-	getResp, err := c.kvClient.GetDeletedSecret(ctx, c.vaultUrl, name, nil)
-	var httpErr *azcore.ResponseError
-	if errors.As(err, &httpErr) {
-		if httpErr.StatusCode != http.StatusNotFound {
-			return nil, err
-		}
+	if options.ResumeToken != "" {
+		return runtime.NewPollerFromResumeToken(
+			options.ResumeToken, c.kvClient.Pipeline(), &runtime.NewPollerFromResumeTokenOptions[DeleteSecretResponse]{Handler: &handler},
+		)
 	}
 
-	return &DeleteSecretPoller{
-		vaultUrl:       c.vaultUrl,
-		secretName:     name,
-		client:         c.kvClient,
-		deleteResponse: delResp,
-		lastResponse:   getResp,
-		resumeToken:    resumeToken,
-	}, nil
+	var rawResp *http.Response
+	ctx = runtime.WithCaptureResponse(ctx, &rawResp)
+	if _, err := c.kvClient.DeleteSecret(ctx, c.vaultUrl, name, options.toGenerated()); err != nil {
+		return nil, err
+	}
+
+	return runtime.NewPoller(rawResp, c.kvClient.Pipeline(), &runtime.NewPollerOptions[DeleteSecretResponse]{Handler: &handler})
 }
 
 // GetDeletedSecretOptions contains optional parameters for GetDeletedSecret.
@@ -363,8 +263,8 @@ type GetDeletedSecretOptions struct {
 	// placeholder for future optional parameters
 }
 
-func (g *GetDeletedSecretOptions) toGenerated() *internal.KeyVaultClientGetDeletedSecretOptions {
-	return &internal.KeyVaultClientGetDeletedSecretOptions{}
+func (g *GetDeletedSecretOptions) toGenerated() *generated.KeyVaultClientGetDeletedSecretOptions {
+	return &generated.KeyVaultClientGetDeletedSecretOptions{}
 }
 
 // GetDeletedSecretResponse is returned by GetDeletedSecret.
@@ -373,7 +273,7 @@ type GetDeletedSecretResponse struct {
 }
 
 // Convert the generated response to the publicly exposed version
-func getDeletedSecretResponseFromGenerated(i internal.KeyVaultClientGetDeletedSecretResponse) GetDeletedSecretResponse {
+func getDeletedSecretResponseFromGenerated(i generated.KeyVaultClientGetDeletedSecretResponse) GetDeletedSecretResponse {
 	vaultURL, name, version := shared.ParseID(i.ID)
 	return GetDeletedSecretResponse{
 		DeletedSecret: DeletedSecret{
@@ -424,7 +324,7 @@ type UpdateSecretPropertiesResponse struct {
 	Secret
 }
 
-func updateSecretPropertiesResponseFromGenerated(i internal.KeyVaultClientUpdateSecretResponse) UpdateSecretPropertiesResponse {
+func updateSecretPropertiesResponseFromGenerated(i generated.KeyVaultClientUpdateSecretResponse) UpdateSecretPropertiesResponse {
 	vaultURL, name, version := shared.ParseID(i.ID)
 	return UpdateSecretPropertiesResponse{
 		Secret: Secret{
@@ -468,7 +368,7 @@ func (c *Client) UpdateSecretProperties(ctx context.Context, secret Secret, opti
 		name,
 		version,
 		secret.toGeneratedProperties(),
-		&internal.KeyVaultClientUpdateSecretOptions{},
+		&generated.KeyVaultClientUpdateSecretOptions{},
 	)
 	if err != nil {
 		return UpdateSecretPropertiesResponse{}, err
@@ -482,8 +382,8 @@ type BackupSecretOptions struct {
 	// placeholder for future optional parameters
 }
 
-func (b *BackupSecretOptions) toGenerated() *internal.KeyVaultClientBackupSecretOptions {
-	return &internal.KeyVaultClientBackupSecretOptions{}
+func (b *BackupSecretOptions) toGenerated() *generated.KeyVaultClientBackupSecretOptions {
+	return &generated.KeyVaultClientBackupSecretOptions{}
 }
 
 // BackupSecretResponse is returned by BackupSecret.
@@ -493,7 +393,7 @@ type BackupSecretResponse struct {
 }
 
 // convert generated response to the publicly exposed version.
-func backupSecretResponseFromGenerated(i internal.KeyVaultClientBackupSecretResponse) BackupSecretResponse {
+func backupSecretResponseFromGenerated(i generated.KeyVaultClientBackupSecretResponse) BackupSecretResponse {
 	return BackupSecretResponse{
 		Value: i.Value,
 	}
@@ -518,8 +418,8 @@ type RestoreSecretBackupOptions struct {
 	// placeholder for future optional parameters
 }
 
-func (r RestoreSecretBackupOptions) toGenerated() *internal.KeyVaultClientRestoreSecretOptions {
-	return &internal.KeyVaultClientRestoreSecretOptions{}
+func (r RestoreSecretBackupOptions) toGenerated() *generated.KeyVaultClientRestoreSecretOptions {
+	return &generated.KeyVaultClientRestoreSecretOptions{}
 }
 
 // RestoreSecretBackupResponse is returned by RestoreSecretBackup.
@@ -528,7 +428,7 @@ type RestoreSecretBackupResponse struct {
 }
 
 // converts the generated response to the publicly exposed version.
-func restoreSecretBackupResponseFromGenerated(i internal.KeyVaultClientRestoreSecretResponse) RestoreSecretBackupResponse {
+func restoreSecretBackupResponseFromGenerated(i generated.KeyVaultClientRestoreSecretResponse) RestoreSecretBackupResponse {
 	vaultURL, name, version := shared.ParseID(i.ID)
 	return RestoreSecretBackupResponse{
 		Secret: Secret{
@@ -562,7 +462,7 @@ func (c *Client) RestoreSecretBackup(ctx context.Context, backup []byte, options
 		options = &RestoreSecretBackupOptions{}
 	}
 
-	resp, err := c.kvClient.RestoreSecret(ctx, c.vaultUrl, internal.SecretRestoreParameters{SecretBundleBackup: backup}, options.toGenerated())
+	resp, err := c.kvClient.RestoreSecret(ctx, c.vaultUrl, generated.SecretRestoreParameters{SecretBundleBackup: backup}, options.toGenerated())
 	if err != nil {
 		return RestoreSecretBackupResponse{}, err
 	}
@@ -575,8 +475,8 @@ type PurgeDeletedSecretOptions struct {
 	// placeholder for future optional parameters
 }
 
-func (p *PurgeDeletedSecretOptions) toGenerated() *internal.KeyVaultClientPurgeDeletedSecretOptions {
-	return &internal.KeyVaultClientPurgeDeletedSecretOptions{}
+func (p *PurgeDeletedSecretOptions) toGenerated() *generated.KeyVaultClientPurgeDeletedSecretOptions {
+	return &generated.KeyVaultClientPurgeDeletedSecretOptions{}
 }
 
 // PurgeDeletedSecretResponse contains the response from method Client.PurgeDeletedSecret.
@@ -585,7 +485,7 @@ type PurgeDeletedSecretResponse struct {
 }
 
 // Converts the generated response to the publicly exposed version.
-func purgeDeletedSecretResponseFromGenerated(i internal.KeyVaultClientPurgeDeletedSecretResponse) PurgeDeletedSecretResponse {
+func purgeDeletedSecretResponseFromGenerated(i generated.KeyVaultClientPurgeDeletedSecretResponse) PurgeDeletedSecretResponse {
 	return PurgeDeletedSecretResponse{}
 }
 
@@ -598,82 +498,15 @@ func (c *Client) PurgeDeletedSecret(ctx context.Context, name string, options *P
 	return purgeDeletedSecretResponseFromGenerated(resp), err
 }
 
-// RecoverDeletedSecretPoller is returned by BeginRecoverDeletedSecret.
-type RecoverDeletedSecretPoller struct {
-	secretName      string
-	vaultUrl        string
-	client          *internal.KeyVaultClient
-	recoverResponse internal.KeyVaultClientRecoverDeletedSecretResponse
-	lastResponse    internal.KeyVaultClientGetSecretResponse
-	rawResponse     *http.Response
-	resumeToken     string
-}
-
-// Done returns true when the polling operation is completed
-func (b *RecoverDeletedSecretPoller) Done() bool {
-	if b.rawResponse == nil {
-		return false
-	}
-	return b.rawResponse.StatusCode == http.StatusOK
-}
-
-// Poll fetches the latest state of the recover operation.
-func (b *RecoverDeletedSecretPoller) Poll(ctx context.Context) (*http.Response, error) {
-	var rawResp *http.Response
-	ctx = runtime.WithCaptureResponse(ctx, &rawResp)
-	resp, err := b.client.GetSecret(ctx, b.vaultUrl, b.secretName, "", nil)
-	if err == nil {
-		// secret has been recovered, finish
-		b.lastResponse = resp
-		b.rawResponse = rawResp
-		return b.rawResponse, nil
-	}
-
-	if rawResp != nil && rawResp.StatusCode == http.StatusNotFound {
-		// this is the expected response
-		b.lastResponse = resp
-		b.rawResponse = rawResp
-		return b.rawResponse, nil
-	}
-
-	return rawResp, err
-}
-
-// FinalResponse returns the final response after the recover operation is complete.
-func (b *RecoverDeletedSecretPoller) FinalResponse(ctx context.Context) (RecoverDeletedSecretResponse, error) {
-	return recoverDeletedSecretResponseFromGenerated(b.recoverResponse), nil
-}
-
-// PollUntilDone polls Key Vault until the recover operation is complete. The t parameter determines the wait between polls.
-func (b *RecoverDeletedSecretPoller) PollUntilDone(ctx context.Context, t time.Duration) (RecoverDeletedSecretResponse, error) {
-	for {
-		resp, err := b.Poll(ctx)
-		if err != nil {
-			b.rawResponse = resp
-		}
-		if b.Done() {
-			break
-		}
-		b.rawResponse = resp
-		time.Sleep(t)
-	}
-	return recoverDeletedSecretResponseFromGenerated(b.recoverResponse), nil
-}
-
-// ResumeToken returns a token for resuming polling at a later time
-func (s *RecoverDeletedSecretPoller) ResumeToken() (string, error) {
-	return s.resumeToken, nil
-}
-
 // BeginRecoverDeletedSecretOptions contains optional parameters for BeginRecoverDeletedSecret.
 type BeginRecoverDeletedSecretOptions struct {
 	// ResumeToken is a string to rehydrate a poller for an operation that has already begun.
-	ResumeToken *string
+	ResumeToken string
 }
 
 // Convert the publicly exposed options object to the generated version
-func (b BeginRecoverDeletedSecretOptions) toGenerated() *internal.KeyVaultClientRecoverDeletedSecretOptions {
-	return &internal.KeyVaultClientRecoverDeletedSecretOptions{}
+func (b BeginRecoverDeletedSecretOptions) toGenerated() *generated.KeyVaultClientRecoverDeletedSecretOptions {
+	return &generated.KeyVaultClientRecoverDeletedSecretOptions{}
 }
 
 // RecoverDeletedSecretResponse is returned by RecoverDeletedSecret.
@@ -682,7 +515,7 @@ type RecoverDeletedSecretResponse struct {
 }
 
 // change recover deleted secret reponse to the generated version.
-func recoverDeletedSecretResponseFromGenerated(i internal.KeyVaultClientRecoverDeletedSecretResponse) RecoverDeletedSecretResponse {
+func recoverDeletedSecretResponseFromGenerated(i generated.KeyVaultClientRecoverDeletedSecretResponse) RecoverDeletedSecretResponse {
 	var a *Properties
 	if i.Attributes != nil {
 		a = &Properties{
@@ -711,48 +544,36 @@ func recoverDeletedSecretResponseFromGenerated(i internal.KeyVaultClientRecoverD
 
 // BeginRecoverDeletedSecret recovers a deleted secret to its latest version. Recovery may take several seconds. This method
 // therefore returns a poller that enables waiting until recovery is complete.
-func (c *Client) BeginRecoverDeletedSecret(ctx context.Context, name string, options *BeginRecoverDeletedSecretOptions) (*RecoverDeletedSecretPoller, error) {
+func (c *Client) BeginRecoverDeletedSecret(ctx context.Context, name string, options *BeginRecoverDeletedSecretOptions) (*runtime.Poller[RecoverDeletedSecretResponse], error) {
 	if options == nil {
 		options = &BeginRecoverDeletedSecretOptions{}
 	}
-	var resumeToken string
-	var recoverResp internal.KeyVaultClientRecoverDeletedSecretResponse
-	var err error
-	if options.ResumeToken == nil {
-		recoverResp, err = c.kvClient.RecoverDeletedSecret(ctx, c.vaultUrl, name, options.toGenerated())
-		if err != nil {
-			return nil, err
-		}
 
-		marshalled, err := json.Marshal(recoverResp)
-		if err != nil {
-			return nil, err
-		}
-		resumeToken = string(marshalled)
-	} else {
-		resumeToken = *options.ResumeToken
-		err = json.Unmarshal([]byte(resumeToken), &recoverResp)
-		if err != nil {
-			return nil, err
-		}
+	handler := beginRecoverDeletedSecretOperation{
+		poll: func(ctx context.Context) (*http.Response, error) {
+			req, err := c.kvClient.GetSecretCreateRequest(ctx, c.vaultUrl, name, "", nil)
+			if err != nil {
+				return nil, err
+			}
+			return c.kvClient.Pipeline().Do(req)
+		},
 	}
 
-	getResp, err := c.kvClient.GetSecret(ctx, c.vaultUrl, name, "", nil)
-	var httpErr *azcore.ResponseError
-	if errors.As(err, &httpErr) {
-		if httpErr.StatusCode != http.StatusNotFound {
-			return nil, err
-		}
+	if options.ResumeToken != "" {
+		return runtime.NewPollerFromResumeToken(
+			options.ResumeToken, c.kvClient.Pipeline(), &runtime.NewPollerFromResumeTokenOptions[RecoverDeletedSecretResponse]{Handler: &handler},
+		)
 	}
 
-	return &RecoverDeletedSecretPoller{
-		lastResponse:    getResp,
-		secretName:      name,
-		client:          c.kvClient,
-		vaultUrl:        c.vaultUrl,
-		recoverResponse: recoverResp,
-		resumeToken:     resumeToken,
-	}, nil
+	var rawResp *http.Response
+	ctx = runtime.WithCaptureResponse(ctx, &rawResp)
+	if _, err := c.kvClient.RecoverDeletedSecret(ctx, c.vaultUrl, name, options.toGenerated()); err != nil {
+		return nil, err
+	}
+
+	return runtime.NewPoller(
+		rawResp, c.kvClient.Pipeline(), &runtime.NewPollerOptions[RecoverDeletedSecretResponse]{Handler: &handler},
+	)
 }
 
 // ListDeletedSecretsResponse contains a page of deleted secrets.
@@ -764,7 +585,7 @@ type ListDeletedSecretsResponse struct {
 	DeletedSecrets []DeletedSecretItem `json:"value,omitempty" azure:"ro"`
 }
 
-func listDeletedSecretsPageFromGenerated(g internal.KeyVaultClientGetDeletedSecretsResponse) ListDeletedSecretsResponse {
+func listDeletedSecretsPageFromGenerated(g generated.KeyVaultClientGetDeletedSecretsResponse) ListDeletedSecretsResponse {
 	var items []DeletedSecretItem
 
 	if len(g.DeletedSecretListResult.Value) > 0 {
@@ -795,14 +616,14 @@ func (c *Client) NewListDeletedSecretsPager(options *ListDeletedSecretsOptions) 
 			var req *policy.Request
 			var err error
 			if page == nil {
-				req, err = c.kvClient.GetDeletedSecretsCreateRequest(ctx, c.vaultUrl, &internal.KeyVaultClientGetDeletedSecretsOptions{})
+				req, err = c.kvClient.GetDeletedSecretsCreateRequest(ctx, c.vaultUrl, &generated.KeyVaultClientGetDeletedSecretsOptions{})
 			} else {
 				req, err = runtime.NewRequest(ctx, http.MethodGet, *page.NextLink)
 			}
 			if err != nil {
 				return ListDeletedSecretsResponse{}, err
 			}
-			resp, err := c.kvClient.Pl.Do(req)
+			resp, err := c.kvClient.Pipeline().Do(req)
 			if err != nil {
 				return ListDeletedSecretsResponse{}, err
 			}
@@ -833,7 +654,7 @@ type ListPropertiesOfSecretVersionsResponse struct {
 }
 
 // create ListSecretsPage from generated pager
-func listSecretVersionsPageFromGenerated(i internal.KeyVaultClientGetSecretVersionsResponse) ListPropertiesOfSecretVersionsResponse {
+func listSecretVersionsPageFromGenerated(i generated.KeyVaultClientGetSecretVersionsResponse) ListPropertiesOfSecretVersionsResponse {
 	var secrets []SecretItem
 	for _, s := range i.Value {
 		secrets = append(secrets, secretItemFromGenerated(s))
@@ -854,14 +675,14 @@ func (c *Client) NewListPropertiesOfSecretVersionsPager(name string, options *Li
 			var req *policy.Request
 			var err error
 			if page == nil {
-				req, err = c.kvClient.GetSecretVersionsCreateRequest(ctx, c.vaultUrl, name, &internal.KeyVaultClientGetSecretVersionsOptions{})
+				req, err = c.kvClient.GetSecretVersionsCreateRequest(ctx, c.vaultUrl, name, &generated.KeyVaultClientGetSecretVersionsOptions{})
 			} else {
 				req, err = runtime.NewRequest(ctx, http.MethodGet, *page.NextLink)
 			}
 			if err != nil {
 				return ListPropertiesOfSecretVersionsResponse{}, err
 			}
-			resp, err := c.kvClient.Pl.Do(req)
+			resp, err := c.kvClient.Pipeline().Do(req)
 			if err != nil {
 				return ListPropertiesOfSecretVersionsResponse{}, err
 			}
@@ -892,7 +713,7 @@ type ListPropertiesOfSecretsResponse struct {
 }
 
 // create a ListSecretsPage from a generated code response
-func listSecretsPageFromGenerated(i internal.KeyVaultClientGetSecretsResponse) ListPropertiesOfSecretsResponse {
+func listSecretsPageFromGenerated(i generated.KeyVaultClientGetSecretsResponse) ListPropertiesOfSecretsResponse {
 	var secrets []SecretItem
 	for _, s := range i.Value {
 		secrets = append(secrets, secretItemFromGenerated(s))
@@ -913,14 +734,14 @@ func (c *Client) NewListPropertiesOfSecretsPager(options *ListPropertiesOfSecret
 			var req *policy.Request
 			var err error
 			if page == nil {
-				req, err = c.kvClient.GetSecretsCreateRequest(ctx, c.vaultUrl, &internal.KeyVaultClientGetSecretsOptions{})
+				req, err = c.kvClient.GetSecretsCreateRequest(ctx, c.vaultUrl, &generated.KeyVaultClientGetSecretsOptions{})
 			} else {
 				req, err = runtime.NewRequest(ctx, http.MethodGet, *page.NextLink)
 			}
 			if err != nil {
 				return ListPropertiesOfSecretsResponse{}, err
 			}
-			resp, err := c.kvClient.Pl.Do(req)
+			resp, err := c.kvClient.Pipeline().Do(req)
 			if err != nil {
 				return ListPropertiesOfSecretsResponse{}, err
 			}