+++ sdk/keyvault/azsecrets/client_test.go
@@ -10,46 +10,15 @@ import (
 	"context"
 	"errors"
 	"fmt"
-	"net/http"
-	"os"
 	"strings"
 	"testing"
 	"time"
 
 	"github.com/Azure/azure-sdk-for-go/sdk/azcore"
-	"github.com/Azure/azure-sdk-for-go/sdk/azcore/log"
 	"github.com/Azure/azure-sdk-for-go/sdk/azcore/to"
-	"github.com/Azure/azure-sdk-for-go/sdk/azidentity"
 	"github.com/Azure/azure-sdk-for-go/sdk/internal/recording"
 	"github.com/stretchr/testify/require"
 )
-
-func TestMain(m *testing.M) {
-	// Initialize
-	if recording.GetRecordMode() == "record" {
-		vaultUrl := os.Getenv("AZURE_KEYVAULT_URL")
-		err := recording.AddURISanitizer("https://fakekvurl.vault.azure.net/", vaultUrl, nil)
-		if err != nil {
-			panic(err)
-		}
-	}
-
-	// Run
-	exitVal := m.Run()
-
-	// cleanup
-
-	os.Exit(exitVal)
-}
-
-func startTest(t *testing.T) func() {
-	err := recording.Start(t, pathToPackage, nil)
-	require.NoError(t, err)
-	return func() {
-		err := recording.Stop(t, nil)
-		require.NoError(t, err)
-	}
-}
 
 func TestSetGetSecret(t *testing.T) {
 	stop := startTest(t)
@@ -63,10 +32,9 @@ func TestSetGetSecret(t *testing.T) {
 	value, err := createRandomName(t, "value")
 	require.NoError(t, err)
 
-	defer cleanUpSecret(t, client, secret)
-
 	_, err = client.SetSecret(context.Background(), secret, value, nil)
 	require.NoError(t, err)
+	defer cleanUpSecret(t, client, secret)
 
 	getResp, err := client.GetSecret(context.Background(), secret, nil)
 	require.NoError(t, err)
@@ -124,17 +92,13 @@ func TestListSecretVersions(t *testing.T) {
 	client, err := createClient(t)
 	require.NoError(t, err)
 
-	secret, err := createRandomName(t, "secret")
-	require.NoError(t, err)
-	value, err := createRandomName(t, "value")
+	secret, err := createRandomName(t, "listsecretversions")
 	require.NoError(t, err)
 
-	_, err = client.SetSecret(context.Background(), secret, value, nil)
-	require.NoError(t, err)
-	_, err = client.SetSecret(context.Background(), secret, value+"1", nil)
-	require.NoError(t, err)
-	_, err = client.SetSecret(context.Background(), secret, value+"2", nil)
-	require.NoError(t, err)
+	for i := 0; i < 3; i++ {
+		_, err = client.SetSecret(context.Background(), secret, fmt.Sprintf("value%d", i), nil)
+		require.NoError(t, err)
+	}
 	defer cleanUpSecret(t, client, secret)
 
 	count := 0
@@ -142,9 +106,13 @@ func TestListSecretVersions(t *testing.T) {
 	for pager.More() {
 		page, err := pager.NextPage(context.Background())
 		require.NoError(t, err)
-		count += len(page.Secrets)
+		for _, secret := range page.Secrets {
+			if strings.HasPrefix(*secret.Name, "listsecretversions") {
+				count++
+			}
+		}
 	}
-	require.GreaterOrEqual(t, count, 3)
+	require.Equal(t, count, 3)
 }
 
 func TestListSecrets(t *testing.T) {
@@ -154,26 +122,24 @@ func TestListSecrets(t *testing.T) {
 	client, err := createClient(t)
 	require.NoError(t, err)
 
-	_, err = client.SetSecret(context.Background(), "secret1", "value", nil)
-	require.NoError(t, err)
-	_, err = client.SetSecret(context.Background(), "secret2", "value", nil)
-	require.NoError(t, err)
-	_, err = client.SetSecret(context.Background(), "secret3", "value", nil)
-	require.NoError(t, err)
-	_, err = client.SetSecret(context.Background(), "secret4", "value", nil)
-	require.NoError(t, err)
-
-	defer cleanUpSecret(t, client, "secret1")
-	defer cleanUpSecret(t, client, "secret2")
-	defer cleanUpSecret(t, client, "secret3")
-	defer cleanUpSecret(t, client, "secret4")
+	for i := 0; i < 4; i++ {
+		name, err := createRandomName(t, fmt.Sprintf("listsecrets%d", i))
+		require.NoError(t, err)
+		_, err = client.SetSecret(context.Background(), name, "value", nil)
+		require.NoError(t, err)
+		defer cleanUpSecret(t, client, name)
+	}
 
 	count := 0
 	pager := client.NewListPropertiesOfSecretsPager(nil)
 	for pager.More() {
 		page, err := pager.NextPage(context.Background())
 		require.NoError(t, err)
-		count += len(page.Secrets)
+		for _, secret := range page.Secrets {
+			if strings.HasPrefix(*secret.Name, "listsecrets") {
+				count++
+			}
+		}
 	}
 	require.Equal(t, count, 4)
 }
@@ -204,12 +170,12 @@ func TestListDeletedSecrets(t *testing.T) {
 	// 2. Delete both secrets
 	resp, err := client.BeginDeleteSecret(context.Background(), secret1, nil)
 	require.NoError(t, err)
-	_, err = resp.PollUntilDone(context.Background(), delay())
+	_, err = resp.PollUntilDone(context.Background(), getPollingOptions())
 	require.NoError(t, err)
 
 	resp, err = client.BeginDeleteSecret(context.Background(), secret2, nil)
 	require.NoError(t, err)
-	_, err = resp.PollUntilDone(context.Background(), delay())
+	_, err = resp.PollUntilDone(context.Background(), getPollingOptions())
 	require.NoError(t, err)
 
 	f := func() {
@@ -225,12 +191,10 @@ func TestListDeletedSecrets(t *testing.T) {
 		secret1: false,
 		secret2: false,
 	}
-	count := 0
 	pager := client.NewListDeletedSecretsPager(nil)
 	for pager.More() {
 		page, err := pager.NextPage(context.Background())
 		require.NoError(t, err)
-		count += len(page.DeletedSecrets)
 		for _, secret := range page.DeletedSecrets {
 			for deleted := range deletedSecrets {
 				if strings.Contains(*secret.ID, deleted) {
@@ -264,7 +228,7 @@ func TestDeleteSecret(t *testing.T) {
 	resp, err := client.BeginDeleteSecret(context.Background(), secret, nil)
 	require.NoError(t, err)
 
-	finalResp, err := resp.PollUntilDone(context.Background(), delay())
+	finalResp, err := resp.PollUntilDone(context.Background(), getPollingOptions())
 	require.NoError(t, err)
 	require.NotNil(t, finalResp.Properties)
 	require.NotNil(t, finalResp.DeletedOn)
@@ -276,12 +240,6 @@ func TestDeleteSecret(t *testing.T) {
 	require.NotNil(t, deleteResp.Properties)
 
 	_, err = client.PurgeDeletedSecret(context.Background(), secret, nil)
-	require.NoError(t, err)
-
-	_, err = client.GetSecret(context.Background(), secret, nil)
-	require.Error(t, err)
-
-	_, err = resp.FinalResponse(context.TODO())
 	require.NoError(t, err)
 }
 
@@ -303,7 +261,7 @@ func TestPurgeDeletedSecret(t *testing.T) {
 	resp, err := client.BeginDeleteSecret(context.Background(), secret, nil)
 	require.NoError(t, err)
 
-	_, err = resp.PollUntilDone(context.Background(), delay())
+	_, err = resp.PollUntilDone(context.Background(), getPollingOptions())
 	require.NoError(t, err)
 
 	_, err = client.PurgeDeletedSecret(context.Background(), secret, nil)
@@ -328,21 +286,17 @@ func TestUpdateSecretProperties(t *testing.T) {
 	client, err := createClient(t)
 	require.NoError(t, err)
 
-	secret, err := createRandomName(t, "secret2")
+	name, err := createRandomName(t, "secret2")
 	require.NoError(t, err)
 	value, err := createRandomName(t, "value")
 	require.NoError(t, err)
 
-	_, err = client.SetSecret(context.Background(), secret, value, nil)
+	setResp, err := client.SetSecret(context.Background(), name, value, nil)
 	require.NoError(t, err)
+	require.Equal(t, *setResp.Secret.Value, value)
+	defer cleanUpSecret(t, client, name)
 
-	defer cleanUpSecret(t, client, secret)
-
-	getResp, err := client.GetSecret(context.Background(), secret, nil)
-	require.NoError(t, err)
-	require.Equal(t, *getResp.Secret.Value, value)
-
-	getResp.Secret.Properties = &Properties{
+	setResp.Secret.Properties = &Properties{
 		ContentType: to.Ptr("password"),
 		Tags: map[string]string{
 			"Tag1": "TagVal1",
@@ -350,13 +304,13 @@ func TestUpdateSecretProperties(t *testing.T) {
 		Enabled:   to.Ptr(true),
 		ExpiresOn: to.Ptr(time.Now().Add(48 * time.Hour)),
 		NotBefore: to.Ptr(time.Now().Add(-24 * time.Hour)),
-		Name:      getResp.Secret.Properties.Name,
+		Name:      setResp.Secret.Properties.Name,
 	}
 
-	_, err = client.UpdateSecretProperties(context.Background(), getResp.Secret, nil)
+	_, err = client.UpdateSecretProperties(context.Background(), setResp.Secret, nil)
 	require.NoError(t, err)
 
-	getResp, err = client.GetSecret(context.Background(), secret, nil)
+	getResp, err := client.GetSecret(context.Background(), name, nil)
 	require.NoError(t, err)
 	require.Equal(t, *getResp.Secret.Value, value)
 	require.Equal(t, getResp.Secret.Properties.Tags["Tag1"], "TagVal1")
@@ -383,16 +337,13 @@ func TestBeginRecoverDeletedSecret(t *testing.T) {
 	pollerResp, err := client.BeginDeleteSecret(context.Background(), secret, nil)
 	require.NoError(t, err)
 
-	_, err = pollerResp.PollUntilDone(context.Background(), delay())
+	_, err = pollerResp.PollUntilDone(context.Background(), getPollingOptions())
 	require.NoError(t, err)
 
 	resp, err := client.BeginRecoverDeletedSecret(context.Background(), secret, nil)
 	require.NoError(t, err)
 
-	_, err = resp.PollUntilDone(context.Background(), delay())
-	require.NoError(t, err)
-
-	_, err = client.SetSecret(context.Background(), secret, value, nil)
+	_, err = resp.PollUntilDone(context.Background(), getPollingOptions())
 	require.NoError(t, err)
 
 	getResp, err := client.GetSecret(context.Background(), secret, nil)
@@ -420,7 +371,7 @@ func TestBeginRecoverDeletedSecretRehydrated(t *testing.T) {
 	pollerResp, err := client.BeginDeleteSecret(context.Background(), secret, nil)
 	require.NoError(t, err)
 
-	_, err = pollerResp.PollUntilDone(context.Background(), delay())
+	_, err = pollerResp.PollUntilDone(context.Background(), getPollingOptions())
 	require.NoError(t, err)
 
 	resp, err := client.BeginRecoverDeletedSecret(context.Background(), secret, nil)
@@ -429,13 +380,10 @@ func TestBeginRecoverDeletedSecretRehydrated(t *testing.T) {
 	rt, err := resp.ResumeToken()
 	require.NoError(t, err)
 
-	poller, err := client.BeginRecoverDeletedSecret(context.Background(), secret, &BeginRecoverDeletedSecretOptions{ResumeToken: &rt})
+	poller, err := client.BeginRecoverDeletedSecret(context.Background(), secret, &BeginRecoverDeletedSecretOptions{ResumeToken: rt})
 	require.NoError(t, err)
 
-	_, err = poller.PollUntilDone(context.Background(), delay())
-	require.NoError(t, err)
-
-	_, err = client.SetSecret(context.Background(), secret, value, nil)
+	_, err = poller.PollUntilDone(context.Background(), getPollingOptions())
 	require.NoError(t, err)
 
 	getResp, err := client.GetSecret(context.Background(), secret, nil)
@@ -450,28 +398,29 @@ func TestBeginDeleteSecretRehydrated(t *testing.T) {
 	client, err := createClient(t)
 	require.NoError(t, err)
 
-	secret, err := createRandomName(t, "begindeleterehydrat")
-	require.NoError(t, err)
-	value, err := createRandomName(t, "value")
+	name, err := createRandomName(t, "begindeleterehydrat")
 	require.NoError(t, err)
 
-	_, err = client.SetSecret(context.Background(), secret, value, nil)
+	_, err = client.SetSecret(context.Background(), name, "value", nil)
 	require.NoError(t, err)
 
-	pollerResp, err := client.BeginDeleteSecret(context.Background(), secret, nil)
+	poller, err := client.BeginDeleteSecret(context.Background(), name, nil)
 	require.NoError(t, err)
 
-	rt, err := pollerResp.ResumeToken()
+	rt, err := poller.ResumeToken()
 	require.NoError(t, err)
 
-	rehydrated, err := client.BeginDeleteSecret(context.Background(), secret, &BeginDeleteSecretOptions{ResumeToken: &rt})
+	rehydrated, err := client.BeginDeleteSecret(context.Background(), name, &BeginDeleteSecretOptions{ResumeToken: rt})
 	require.NoError(t, err)
 
-	_, err = rehydrated.PollUntilDone(context.Background(), delay())
+	_, err = rehydrated.PollUntilDone(context.Background(), getPollingOptions())
 	require.NoError(t, err)
 
-	_, err = client.GetSecret(context.Background(), secret, nil)
-	require.Error(t, err)
+	_, err = client.GetDeletedSecret(context.Background(), name, nil)
+	require.NoError(t, err)
+
+	_, err = client.PurgeDeletedSecret(context.Background(), name, nil)
+	require.NoError(t, err)
 }
 
 func TestBackupSecret(t *testing.T) {
@@ -481,98 +430,45 @@ func TestBackupSecret(t *testing.T) {
 	client, err := createClient(t)
 	require.NoError(t, err)
 
-	secret, err := createRandomName(t, "secrets")
+	name, err := createRandomName(t, "testbackupsecret")
 	require.NoError(t, err)
 	value, err := createRandomName(t, "value")
 	require.NoError(t, err)
 
-	_, err = client.SetSecret(context.Background(), secret, value, nil)
+	_, err = client.SetSecret(context.Background(), name, value, nil)
 	require.NoError(t, err)
 
-	defer cleanUpSecret(t, client, secret)
+	defer cleanUpSecret(t, client, name)
 
-	backupResp, err := client.BackupSecret(context.Background(), secret, nil)
+	backupResp, err := client.BackupSecret(context.Background(), name, nil)
 	require.NoError(t, err)
 	require.Greater(t, len(backupResp.Value), 0)
 
-	respPoller, err := client.BeginDeleteSecret(context.Background(), secret, nil)
+	respPoller, err := client.BeginDeleteSecret(context.Background(), name, nil)
 	require.NoError(t, err)
-	_, err = respPoller.PollUntilDone(context.Background(), delay())
-	require.NoError(t, err)
-
-	_, err = client.PurgeDeletedSecret(context.Background(), secret, nil)
+	_, err = respPoller.PollUntilDone(context.Background(), getPollingOptions())
 	require.NoError(t, err)
 
-	_, err = client.GetSecret(context.Background(), secret, nil)
-	var httpErr *azcore.ResponseError
-	require.True(t, errors.As(err, &httpErr))
-	require.Equal(t, httpErr.RawResponse.StatusCode, http.StatusNotFound)
+	_, err = client.PurgeDeletedSecret(context.Background(), name, nil)
+	require.NoError(t, err)
 
-	_, err = client.GetDeletedSecret(context.Background(), secret, nil)
-	require.True(t, errors.As(err, &httpErr))
-	require.Equal(t, httpErr.RawResponse.StatusCode, http.StatusNotFound)
-
-	time.Sleep(20 * delay())
-
-	// Poll this operation manually
 	var restoreResp RestoreSecretBackupResponse
-	var i int
-	for i = 0; i < 20; i++ {
+	for i := 0; i < 20; i++ {
 		restoreResp, err = client.RestoreSecretBackup(context.Background(), backupResp.Value, nil)
 		if err == nil {
+			require.Equal(t, *restoreResp.Name, name)
+			require.NotNil(t, restoreResp.ID)
 			break
 		}
-		time.Sleep(delay())
+		var respErr *azcore.ResponseError
+		if errors.As(err, &respErr) {
+			// Key Vault will return 409 until it completes the purge
+			if respErr.RawResponse.StatusCode != 409 {
+				require.NoError(t, err)
+			}
+		} else {
+			require.NoError(t, err)
+		}
+		recording.Sleep(30 * time.Second)
 	}
-	require.NoError(t, err)
-	require.Contains(t, *restoreResp.ID, secret)
-
-	// Now the Secret should be Get-able
-	_, err = client.GetSecret(context.Background(), secret, nil)
-	require.NoError(t, err)
-}
-
-func TestTimeout(t *testing.T) {
-	fakeKVUrl := "https://test-sync-time-dummy.vault.azure.net/"
-	cred, err := azidentity.NewDefaultAzureCredential(nil)
-	require.NoError(t, err)
-
-	client, err := NewClient(fakeKVUrl, cred, nil)
-	require.NoError(t, err)
-
-	c := context.Background()
-	c, cancelFunc := context.WithTimeout(c, 10*time.Second)
-	defer cancelFunc()
-
-	start := time.Now()
-	_, err = client.GetSecret(c, "nonexistentsecret", nil)
-	require.Error(t, err)
-	require.ErrorIs(t, err, context.DeadlineExceeded)
-	require.Less(t, time.Since(start).Seconds(), 11.0)
-	require.Greater(t, time.Since(start).Seconds(), 9.0)
-}
-
-func TestLogging(t *testing.T) {
-	fakeKVUrl := "https://test-sync-time-dummy.vault.azure.net/"
-	cred, err := azidentity.NewDefaultAzureCredential(nil)
-	require.NoError(t, err)
-
-	client, err := NewClient(fakeKVUrl, cred, nil)
-	require.NoError(t, err)
-
-	c := context.Background()
-	c, cancelFunc := context.WithTimeout(c, 10*time.Second)
-	defer cancelFunc()
-
-	log.SetListener(func(cls log.Event, msg string) {
-		fmt.Println(msg)
-	})
-	log.SetEvents(log.EventRequest, log.EventResponse)
-
-	start := time.Now()
-	_, err = client.GetSecret(c, "nonexistentsecret", nil)
-	require.Error(t, err)
-	require.ErrorIs(t, err, context.DeadlineExceeded)
-	require.Less(t, time.Since(start).Seconds(), 11.0)
-	require.Greater(t, time.Since(start).Seconds(), 9.0)
 }