+++ sdk/keyvault/azsecrets/utils_test.go
@@ -11,17 +11,61 @@ import (
 	"fmt"
 	"hash/fnv"
 	"os"
+	"strings"
 	"testing"
 	"time"
 
 	"github.com/Azure/azure-sdk-for-go/sdk/azcore"
 	"github.com/Azure/azure-sdk-for-go/sdk/azcore/policy"
+	"github.com/Azure/azure-sdk-for-go/sdk/azcore/runtime"
 	"github.com/Azure/azure-sdk-for-go/sdk/azidentity"
 	"github.com/Azure/azure-sdk-for-go/sdk/internal/recording"
 	"github.com/stretchr/testify/require"
 )
 
-var pathToPackage = "sdk/keyvault/azsecrets/testdata"
+const (
+	fakeVaultURL  = "https://fakekvurl.vault.azure.net"
+	pathToPackage = "sdk/keyvault/azsecrets/testdata"
+)
+
+var liveVaultURL string
+
+func TestMain(m *testing.M) {
+	liveVaultURL = strings.TrimSuffix(os.Getenv("AZURE_KEYVAULT_URL"), "/")
+	if liveVaultURL == "" && recording.GetRecordMode() != recording.PlaybackMode {
+		panic("no value for AZURE_KEYVAULT_URL")
+	}
+	err := recording.ResetProxy(nil)
+	if err != nil {
+		panic(err)
+	}
+	if recording.GetRecordMode() == recording.RecordingMode {
+		err := recording.AddURISanitizer(fakeVaultURL, liveVaultURL, nil)
+		if err != nil {
+			panic(err)
+		}
+		err = recording.AddBodyRegexSanitizer(fakeVaultURL, liveVaultURL, nil)
+		if err != nil {
+			panic(err)
+		}
+		defer func() {
+			err := recording.ResetProxy(nil)
+			if err != nil {
+				panic(err)
+			}
+		}()
+	}
+	os.Exit(m.Run())
+}
+
+func startTest(t *testing.T) func() {
+	err := recording.Start(t, pathToPackage, nil)
+	require.NoError(t, err)
+	return func() {
+		err := recording.Stop(t, nil)
+		require.NoError(t, err)
+	}
+}
 
 func createRandomName(t *testing.T, prefix string) (string, error) {
 	h := fnv.New32a()
@@ -38,7 +82,10 @@ func lookupEnvVar(s string) string {
 }
 
 func createClient(t *testing.T) (*Client, error) {
-	vaultUrl := recording.GetEnvVariable("AZURE_KEYVAULT_URL", "https://fakekvurl.vault.azure.net/")
+	vaultURL := liveVaultURL
+	if vaultURL == "" {
+		vaultURL = fakeVaultURL
+	}
 
 	client, err := recording.NewRecordingHTTPClient(t, nil)
 	require.NoError(t, err)
@@ -57,45 +104,37 @@ func createClient(t *testing.T) (*Client, error) {
 		cred, err = azidentity.NewClientSecretCredential(tenantId, clientId, clientSecret, nil)
 		require.NoError(t, err)
 	} else {
-		cred = NewFakeCredential("fake", "fake")
+		cred = NewFakeCredential()
 	}
 
-	return NewClient(vaultUrl, cred, options)
+	return NewClient(vaultURL, cred, options)
 }
 
-func delay() time.Duration {
-	if recording.GetRecordMode() == "playback" {
-		return 1 * time.Microsecond
+func getPollingOptions() *runtime.PollUntilDoneOptions {
+	freq := time.Second
+	if recording.GetRecordMode() == recording.RecordingMode {
+		freq = time.Minute
 	}
-	return 250 * time.Millisecond
+	return &runtime.PollUntilDoneOptions{Frequency: freq}
 }
 
 func cleanUpSecret(t *testing.T, client *Client, secret string) {
 	resp, err := client.BeginDeleteSecret(context.Background(), secret, nil)
 	require.NoError(t, err)
 
-	_, err = resp.PollUntilDone(context.Background(), delay())
+	_, err = resp.PollUntilDone(context.Background(), getPollingOptions())
 	require.NoError(t, err)
 
 	_, err = client.PurgeDeletedSecret(context.Background(), secret, nil)
 	require.NoError(t, err)
 }
 
-type FakeCredential struct {
-	accountName string
-	accountKey  string
-}
+type FakeCredential struct{}
 
-func NewFakeCredential(accountName, accountKey string) *FakeCredential {
-	return &FakeCredential{
-		accountName: accountName,
-		accountKey:  accountKey,
-	}
+func NewFakeCredential() *FakeCredential {
+	return &FakeCredential{}
 }
 
 func (f *FakeCredential) GetToken(ctx context.Context, options policy.TokenRequestOptions) (azcore.AccessToken, error) {
-	return azcore.AccessToken{
-		Token:     "faketoken",
-		ExpiresOn: time.Date(2040, time.January, 1, 1, 1, 1, 1, time.UTC),
-	}, nil
+	return azcore.AccessToken{Token: "faketoken", ExpiresOn: time.Now().Add(time.Hour).UTC()}, nil
 }